move `OutputObjectDefinitoin` from private `_output.py` to public `output.py`
